Move nsidc-0082 fetch code into amsr submodule
